fix: Reorder the code to correct files
